requirements.txt
python-telegram-bot
logging
pandas
SpeechRecognition
pydub
translate
nest_asyncio
dotenv
